Minor bug fixes.
Fixed a bug where the data wouldn't be able to get put because the server url wasn't set as a put method, but a get method instead.

There was a problem with the post entry file that caused an error when editing the form due to values being set as null or undefined. States are set to update when the prop changes, not sure if there is a better way to do this?