Reply & delete comments. Collapse threads & breadcrumbs
Added minor functionality to reply to comments. The backend needs work as it is currently not possible and I am not sure how to make it work.

Added ability to delete comments. Currently anyone can delete any comment so it will need authentication.

Comment threads are collapsible. Single comments are also collapsible. When collapsed they display a compact view of the comment.

Added a very simple breadcrumbs to the top of the view post thread. This can be made to be more complex if truly needed.