Modify comment request
Modified the comment request so that the comment data is received completely on request for the post in the correct nesting order.

Added a backend function that will nest the comment with their parent comments recursively.

Modified the get all comments to get all the comments from one post and then that is given to the nesting function to return the nested data to the client.