Adjustments required
Added several adjustments, most of which came from Jaden's comment on the previous pr. These adjustments fix the functionality of the app.

Added a create button on the home page so that user can click it and it will link them to the create post page.

Posts on the homepage can now be clicked so that users can view them.

Viewed posts properly show their usernames and profile pictures under the title. The same for comments.

Images were broken because of them not properly being implemented. Changed this to properly implement images by creating an assets and images folder for them alongside importing the image where needed.

Posts can now be deleted.

Imported plugin for tailwind to limit the line count on the homepage posts. Alongside only allowing users that own the post to edit/delete.

Fixed warnings in console.