Visual Adjustments
Applied visual adjustments so that this page follows the design standard.

Removed the separate search queries and included the search text from the header search bar. The search now uses the header search instead of the three separate responses.

Modified search query in server side to include or operator in the where clause of the admin search.