Add comment editing.
Added comment editing.

In the comment dropdown menu there is a new button to edit the comment. This can be clicked and, similar to a reply, it will put the reference comment above the add comment input. The comment text will then be prefilled into the input where the user can edit the text. The reference comment has a yellow border when editing to distinguish between a reply or a edited comment.

Comments are ordered by created date instead of updated data as that broke the nesting when editing the comments.

Small bug fixes and visual fixes.